requirements.txt
matplotlib==3.6.0
numpy==1.23.1
PyYAML==6.0
torch==1.11.0
torchvision==0.12.0
